# public/openapi.yaml
openapi: 3.0.1
info:
  title: TODO Plugin
  description: A plugin that allows the user to create and manage a TODO list using ChatGPT.
  version: '0.3.1'
servers:
  - url: https://magong.se
    description: Production server
paths:
  /api/todos:
    get:
      operationId: getTodoList
      summary: Get the list of todos or a specific todo
      parameters:
        - name: id
          in: query
          required: false
          schema:
            type: number
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/getTodosResponse'
    post:
      operationId: addOneTodo
      summary: Add a todo, without an id
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/addTodoRequest'
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/getTodosResponse'
    delete:
      operationId: deleteOneTodo
      summary: Delete a todo by id
      parameters:
        - name: id
          in: query
          required: true
          schema:
            type: number
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/getTodosResponse'
    put:
      operationId: updateOneTodo
      summary: Update a todo by id
      parameters:
        - name: id
          in: query
          required: true
          schema:
            type: number
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/updateTodoRequest'
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/getTodosResponse'
components:
  schemas:
    getTodosResponse:
      type: object
      properties:
        todos:
          type: array
          items:
            type: object
            description: A todo
            properties:
              id:
                type: number
              msg:
                type: string
    addTodoRequest:
      type: object
      description: A todo without an id
      properties:
        msg:
          type: string
          required: true
    updateTodoRequest:
      type: object
      description: A full todo
      properties:
        id:
          type: number
          required: true
        msg:
          type: string
          required: true